+++ src/main/java/com/corinne/corinne_be/service/PriceService.java
@@ -133,7 +133,7 @@ public class PriceService {
         for(DateCandle dateCandle : dateCandles){
 
             // ---> 임의로 넣은 현재가 가격 현재가 수정 필수
-            Long currentPrice = redisRepository.getTradePrice(dateCandle.getTiker());
+            int currentPrice = redisRepository.getTradePrice(dateCandle.getTiker());
 
             String tiker = dateCandle.getTiker();
